minimega: lots of refactoring for namespaces.
Namespaces seems to be working better than ever and it's less hacky than
ever... great! We can even include the head node in the namespace
without things deadlocking.
