+++ src/minimega/command_meshage.go
@@ -89,7 +89,7 @@ vm info from nodes kn1 and kn2:
 
 You can use 'all' to send a command to all connected clients.`,
 		Patterns: []string{
-			"mesh send <vms or all> (command)",
+			"mesh send <clients or all> (command)",
 		},
 		Call: cliMeshageSend,
 	},
@@ -107,6 +107,13 @@ func meshageHandler() {
 				return
 			}
 			cmd.SetSource(SourceMeshage)
+
+			// Copy the Record flag at each level of nested command
+			for c, c2 := cmd, &mCmd.Command; c != nil && c2 != nil; {
+				// could all be the post statement...
+				c.Record = c2.Record
+				c, c2 = c.Subcommand, c2.Subcommand
+			}
 
 			resps := []minicli.Responses{}
 			for resp := range runCommand(cmd) {
@@ -248,5 +255,5 @@ func cliMeshageTimeout(c *minicli.Command) *minicli.Response {
 }
 
 func cliMeshageSend(c *minicli.Command, respChan chan minicli.Responses) {
-	meshageSend(c.Subcommand, c.StringArgs["vms"], respChan)
+	meshageSend(c.Subcommand, c.StringArgs["clients"], respChan)
 }
