+++ src/minimega/vmconfig_cli.go
@@ -1,0 +1,552 @@
+// Copyright (2015) Sandia Corporation.
+// Under the terms of Contract DE-AC04-94AL85000 with Sandia Corporation,
+// the U.S. Government retains certain rights in this software.
+
+package main
+
+import (
+	"fmt"
+	"log"
+	"minicli"
+)
+
+var vmconfigCLIHandlers = []minicli.Handler{
+	{ // vm config
+		HelpShort: "display, save, or restore the current VM configuration",
+		HelpLong: `
+Display, save, or restore the current VM configuration. Note that saving and
+restoring configuration applies to all VM configurations including KVM-based VM
+configurations.
+
+To display the current configuration, call vm config with no arguments.
+
+List the current saved configurations with 'vm config restore'.
+
+To save a configuration:
+
+	vm config save <config name>
+
+To restore a configuration:
+
+	vm config restore <config name>
+
+To clone the configuration of an existing VM:
+
+	vm config clone <vm name or id>
+
+Calling clear vm config will clear all VM configuration options, but will not
+remove saved configurations.`,
+		Patterns: []string{
+			"vm config",
+			"vm config <save,> <name>",
+			"vm config <restore,> [name]",
+			"vm config <clone,> <vm id or name>",
+		},
+		Call: wrapSimpleCLI(cliVmConfig),
+	},
+	{ // vm config memory
+		HelpShort: "set the amount of physical memory for a VM",
+		HelpLong: `
+Set the amount of physical memory to allocate in megabytes.`,
+		Patterns: []string{
+			"vm config memory [memory in megabytes]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "memory")
+		}),
+	},
+	{ // vm config vcpus
+		HelpShort: "set the number of virtual CPUs for a VM",
+		HelpLong: `
+Set the number of virtual CPUs to allocate for a VM.`,
+		Patterns: []string{
+			"vm config vcpus [number of CPUs]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "vcpus")
+		}),
+	},
+	{ // vm config cpus
+		HelpShort: "set the virtual CPU architecture",
+		HelpLong: `
+Set the virtual CPU architecture.
+
+By default, set to 'host' which matches the host architecture. See 'kvm -cpu
+help' for a list of architectures available for your version of kvm.`,
+		Patterns: []string{
+			"vm config cpu [cpu]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "cpu")
+		}),
+	},
+	{ // vm config net
+		HelpShort: "specific the networks a VM is a member of",
+		HelpLong: `
+Specify the network(s) that the VM is a member of by VLAN. A corresponding VLAN
+will be created for each network. Optionally, you may specify the bridge the
+interface will be connected on. If the bridge name is omitted, minimega will
+use the default 'mega_bridge'. You can also optionally specify the mac address
+of the interface to connect to that network. If not specifed, the mac address
+will be randomly generated. Additionally, you can optionally specify a driver
+for qemu to use. By default, e1000 is used.
+
+Examples:
+
+To connect a VM to VLANs 1 and 5:
+
+	vm config net 1 5
+
+To connect a VM to VLANs 100, 101, and 102 with specific mac addresses:
+
+	vm config net 100,00:00:00:00:00:00 101,00:00:00:00:01:00 102,00:00:00:00:02:00
+
+To connect a VM to VLAN 1 on bridge0 and VLAN 2 on bridge1:
+
+	vm config net bridge0,1 bridge1,2
+
+To connect a VM to VLAN 100 on bridge0 with a specific mac:
+
+	vm config net bridge0,100,00:11:22:33:44:55
+
+To specify a specific driver, such as i82559c:
+
+	vm config net 100,i82559c
+
+If you prefer, you can also use aliases for VLANs:
+
+	vm config net DMZ CORE
+
+These aliases will be allocated from the pool of available VLANs and is
+namespace-aware (i.e. 'DMZ' in namespace 'foo' will be a different VLAN than
+'DMZ' in namespace 'foo'). Internally, this is implemented by concatenating the
+namespace name with the VLAN alias (e.g. 'DMZ' in namespace 'foo' becomes
+'foo.DMZ'). If you wish to connect VLANs in different namespaces, you may use
+abuse this implementation detail:
+
+	namespace bar
+	vm config net foo.DMZ
+
+Calling vm net with no parameters will list the current networks for this VM.`,
+		Patterns: []string{
+			"vm config net [netspec]...",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "net")
+		}),
+	},
+	{ // vm config append
+		HelpShort: "set an append string to pass to a kernel set with vm kernel",
+		HelpLong: `
+Add an append string to a kernel set with vm kernel. Setting vm append without
+using vm kernel will result in an error.
+
+For example, to set a static IP for a linux VM:
+
+	vm kvm config append ip=10.0.0.5 gateway=10.0.0.1 netmask=255.255.255.0 dns=10.10.10.10
+
+Note: this configuration only applies to KVM-based VMs.`,
+		Patterns: []string{
+			"vm config append [arg]...",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "append")
+		}),
+	},
+	{ // vm config qemu
+		HelpShort: "set the QEMU process to invoke. Relative paths are ok.",
+		HelpLong: `
+Set the QEMU process to invoke. Relative paths are ok.
+
+Note: this configuration only applies to KVM-based VMs.`,
+		Patterns: []string{
+			"vm config qemu [path to qemu]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "qemu")
+		}),
+	},
+	{ // vm config qemu-override
+		HelpShort: "override parts of the QEMU launch string",
+		HelpLong: `
+Override parts of the QEMU launch string by supplying a string to match, and a
+replacement string.
+
+Note: this configuration only applies to KVM-based VMs.`,
+		Patterns: []string{
+			"vm config qemu-override",
+			"vm config qemu-override add <match> <replacement>",
+			"vm config qemu-override delete <id or all>",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "qemu-override")
+		}),
+	},
+	{ // vm config qemu-append
+		HelpShort: "add additional arguments to the QEMU command",
+		HelpLong: `
+Add additional arguments to be passed to the QEMU instance. For example:
+
+	vm kvm config qemu-append -serial tcp:localhost:4001
+
+Note: this configuration only applies to KVM-based VMs.`,
+		Patterns: []string{
+			"vm config qemu-append [argument]...",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "qemu-append")
+		}),
+	},
+	{ // vm config migrate
+		HelpShort: "set migration image for a saved VM",
+		HelpLong: `
+Assign a migration image, generated by a previously saved VM to boot with.
+Migration images should be booted with a kernel/initrd, disk, or cdrom. Use 'vm
+migrate' to generate migration images from running VMs.
+
+Note: this configuration only applies to KVM-based VMs.`,
+		Patterns: []string{
+			"vm config migrate [path to migration image]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "migrate")
+		}),
+	},
+	{ // vm config disk
+		HelpShort: "set disk images to attach to a VM",
+		HelpLong: `
+Attach one or more disks to a vm. Any disk image supported by QEMU is a valid
+parameter. Disk images launched in snapshot mode may safely be used for
+multiple VMs.
+
+Note: this configuration only applies to KVM-based VMs.`,
+		Patterns: []string{
+			"vm config disk [path to disk image]...",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "disk")
+		}),
+	},
+	{ // vm config cdrom
+		HelpShort: "set a cdrom image to attach to a VM",
+		HelpLong: `
+Attach a cdrom to a VM. When using a cdrom, it will automatically be set to be
+the boot device.
+
+Note: this configuration only applies to KVM-based VMs.`,
+		Patterns: []string{
+			"vm config cdrom [path to cdrom image]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "cdrom")
+		}),
+	},
+	{ // vm config kernel
+		HelpShort: "set a kernel image to attach to a VM",
+		HelpLong: `
+Attach a kernel image to a VM. If set, QEMU will boot from this image instead
+of any disk image.
+
+Note: this configuration only applies to KVM-based VMs.`,
+		Patterns: []string{
+			"vm config kernel [path to kernel]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "kernel")
+		}),
+	},
+	{ // vm config initrd
+		HelpShort: "set a initrd image to attach to a VM",
+		HelpLong: `
+Attach an initrd image to a VM. Passed along with the kernel image at boot
+time.
+
+Note: this configuration only applies to KVM-based VMs.`,
+		Patterns: []string{
+			"vm config initrd [path to initrd]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "initrd")
+		}),
+	},
+	{ // vm config uuid
+		HelpShort: "set the UUID for a VM",
+		HelpLong: `
+Set the UUID for a virtual machine. If not set, minimega will create a random
+one when the VM is launched.
+
+Note: this configuration only applies to KVM-based VMs.`,
+		Patterns: []string{
+			"vm config uuid [uuid]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "uuid")
+		}),
+	},
+	{ // vm config serial
+		HelpShort: "specify the serial ports a VM will use",
+		HelpLong: `
+Specify the serial ports that will be created for the VM to use.
+Serial ports specified will be mapped to the VM's /dev/ttySX device, where X
+refers to the connected unix socket on the host at
+$minimega_runtime/<vm_id>/serialX.
+
+Examples:
+
+To display current serial ports:
+  vm config serial
+
+To create three serial ports:
+  vm config serial 3
+
+Note: Whereas modern versions of Windows support up to 256 COM ports, Linux
+typically only supports up to four serial devices. To use more, make sure to
+pass "8250.n_uarts = 4" to the guest Linux kernel at boot. Replace 4 with
+another number.`,
+		Patterns: []string{
+			"vm config serial [number of serial ports]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "serial")
+		}),
+	},
+	{ // vm config virtio-serial
+		HelpShort: "specify the virtio-serial ports a VM will use",
+		HelpLong: `
+Specify the virtio-serial ports that will be created for the VM to use.
+Virtio-serial ports specified will be mapped to the VM's
+/dev/virtio-port/<portname> device, where <portname> refers to the connected
+unix socket on the host at $minimega_runtime/<vm_id>/virtio-serialX.
+
+Examples:
+
+To display current virtio-serial ports:
+  vm config virtio-serial
+
+To create three virtio-serial ports:
+  vm config virtio-serial 3`,
+		Patterns: []string{
+			"vm config virtio-serial [number of virtio-serial ports]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "virtio-serial")
+		}),
+	},
+	{ // vm config snapshot
+		HelpShort: "enable or disable snapshot mode when using disk images",
+		HelpLong: `
+Enable or disable snapshot mode when using disk images. When enabled, disks
+images will be loaded in memory when run and changes will not be saved. This
+allows a single disk image to be used for many VMs.
+
+Note: this configuration only applies to KVM-based VMs.`,
+		Patterns: []string{
+			"vm config snapshot [true,false]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "snapshot")
+		}),
+	},
+	{ // vm config hostname
+		HelpShort: "set a hostname for containers",
+		HelpLong: `
+Set a hostname for a container before launching the init program. If not set,
+the hostname will be that of the physical host. The hostname can also be set by
+the init program or other root process in the container.`,
+		Patterns: []string{
+			"vm config hostname [hostname]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "hostname")
+		}),
+	},
+	{ // vm config init
+		HelpShort: "container init program and args",
+		HelpLong: `
+Set the init program and args to exec into upon container launch. This will be
+PID 1 in the container.`,
+		Patterns: []string{
+			"vm config init [init]...",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "init")
+		}),
+	},
+	{ // vm config filesystem
+		HelpShort: "set the filesystem for containers",
+		HelpLong: `
+Set the filesystem to use for launching a container. This should be a root
+filesystem for a linux distribution (containing /dev, /proc, /sys, etc.)`,
+		Patterns: []string{
+			"vm config filesystem [filesystem]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "filesystem")
+		}),
+	},
+	{ // vm config fifo
+		HelpShort: "set the number of fifos for containers",
+		HelpLong: `
+Set the number of named pipes to include in the container for container-host
+communication. Named pipes will appear on the host in the instance directory
+for the container as fifoN, and on the container as /dev/fifos/fifoN.
+
+Fifos are created using mkfifo() and have all of the same usage constraints.`,
+		Patterns: []string{
+			"vm config fifo [number]",
+		},
+		Call: wrapSimpleCLI(func(c *minicli.Command) *minicli.Response {
+			return cliVmConfigField(c, "fifo")
+		}),
+	},
+	{ // clear vm config
+		HelpShort: "reset vm config to the default value",
+		HelpLong: `
+Resets the configuration for a provided field (or the whole configuration) back
+to the default value.`,
+		// HACK: These patterns could be reduced to a single pattern with all
+		// the different config fields as one multiple choice, however, to make
+		// it easier to read, we split them into separate patterns. We could
+		// use string literals for the field names but then we'd have to
+		// process the Original string within the Command struct to figure out
+		// what field we're supposed to clear. Instead, we can leverage the
+		// magic of single-choice fields to set the field name in BoolArgs.
+		Patterns: []string{
+			"clear vm config",
+			// VMConfig
+			"clear vm config <namespace,>",
+			"clear vm config <memory,>",
+			"clear vm config <net,>",
+			"clear vm config <vcpus,>",
+			// KVMConfig
+			"clear vm config <append,>",
+			"clear vm config <cdrom,>",
+			"clear vm config <migrate,>",
+			"clear vm config <disk,>",
+			"clear vm config <initrd,>",
+			"clear vm config <kernel,>",
+			"clear vm config <qemu,>",
+			"clear vm config <qemu-append,>",
+			"clear vm config <qemu-override,>",
+			"clear vm config <snapshot,>",
+			"clear vm config <uuid,>",
+			"clear vm config <serial,>",
+			"clear vm config <virtio-serial,>",
+			// ContainerConfig
+			"clear vm config <hostname,>",
+			"clear vm config <filesystem,>",
+			"clear vm config <init,>",
+		},
+		Call: wrapSimpleCLI(cliClearVmConfig),
+	},
+}
+
+func cliVmConfig(c *minicli.Command) *minicli.Response {
+	resp := &minicli.Response{Host: hostname}
+
+	if c.BoolArgs["save"] {
+		// Save the current config
+		savedInfo[c.StringArgs["name"]] = *vmConfig.Copy()
+	} else if c.BoolArgs["restore"] {
+		if name, ok := c.StringArgs["name"]; ok {
+			// Try to restore an existing config
+			if s, ok := savedInfo[name]; ok {
+				vmConfig = *s.Copy()
+			} else {
+				resp.Error = fmt.Sprintf("config %v does not exist", name)
+			}
+		} else if len(savedInfo) == 0 {
+			resp.Error = "no vm configs saved"
+		} else {
+			// List the save configs
+			for k := range savedInfo {
+				resp.Response += fmt.Sprintln(k)
+			}
+		}
+	} else if c.BoolArgs["clone"] {
+		// Clone the config of an existing vm
+		vm := LocalVMs().findVm(c.StringArgs["vm"])
+		if vm == nil {
+			resp.Error = vmNotFound(c.StringArgs["vm"]).Error()
+		} else {
+			vmConfig.BaseConfig = *vm.Config().Copy()
+			switch vm := vm.(type) {
+			case *KvmVM:
+				vmConfig.KVMConfig = *vm.KVMConfig.Copy()
+			case *ContainerVM:
+				vmConfig.ContainerConfig = *vm.ContainerConfig.Copy()
+			}
+		}
+	} else {
+		// Print the config
+		resp.Response = vmConfig.String()
+	}
+
+	return resp
+}
+
+func cliVmConfigField(c *minicli.Command, field string) *minicli.Response {
+	resp := &minicli.Response{Host: hostname}
+
+	// If there are no args it means that we want to display the current value
+	nArgs := len(c.StringArgs) + len(c.ListArgs) + len(c.BoolArgs)
+
+	var ok bool
+	var fns VMConfigFns
+	var config interface{}
+
+	// Find the right config functions, baseConfigFns has highest priority
+	if fns, ok = baseConfigFns[field]; ok {
+		config = &vmConfig.BaseConfig
+	} else if fns, ok = kvmConfigFns[field]; ok {
+		config = &vmConfig.KVMConfig
+	} else if fns, ok = containerConfigFns[field]; ok {
+		config = &vmConfig.ContainerConfig
+	} else {
+		log.Fatal("unknown config field: `%s`", field)
+	}
+
+	if nArgs == 0 {
+		resp.Response = fns.Print(config)
+	} else {
+		if err := fns.Update(config, c); err != nil {
+			resp.Error = err.Error()
+		}
+	}
+
+	return resp
+}
+
+func cliClearVmConfig(c *minicli.Command) *minicli.Response {
+	resp := &minicli.Response{Host: hostname}
+
+	var clearAll = len(c.BoolArgs) == 0
+	var clearKVM = clearAll || (len(c.BoolArgs) == 1 && c.BoolArgs["kvm"])
+	var clearContainer = clearAll || (len(c.BoolArgs) == 1 && c.BoolArgs["container"])
+	var cleared bool
+
+	for k, fns := range baseConfigFns {
+		if clearAll || c.BoolArgs[k] {
+			fns.Clear(&vmConfig.BaseConfig)
+			cleared = true
+		}
+	}
+	for k, fns := range kvmConfigFns {
+		if clearKVM || c.BoolArgs[k] {
+			fns.Clear(&vmConfig.KVMConfig)
+			cleared = true
+		}
+	}
+	for k, fns := range containerConfigFns {
+		if clearContainer || c.BoolArgs[k] {
+			fns.Clear(&vmConfig.ContainerConfig)
+			cleared = true
+		}
+	}
+
+	if !cleared {
+		log.Fatalln("no callback defined for clear")
+	}
+
+	return resp
+}
