+++ src/minimega/web.go
@@ -150,11 +150,11 @@ func webScreenshot(w http.ResponseWriter, r *http.Request) {
 
 	cmdStr := fmt.Sprintf("vm screenshot %s file /dev/null %s", id, size)
 	if host != hostname {
-		cmdStr = fmt.Sprintf("mesh send %s .record false %s", host, cmdStr)
+		cmdStr = fmt.Sprintf("mesh send %s %s", host, cmdStr)
 	}
 
 	cmd := minicli.MustCompile(cmdStr)
-	cmd.Record = false
+	cmd.SetRecord(false)
 
 	var screenshot []byte
 
@@ -208,7 +208,7 @@ func webHosts(w http.ResponseWriter, r *http.Request) {
 	hosts := [][]interface{}{}
 
 	cmd := minicli.MustCompile("host")
-	cmd.Record = false
+	cmd.SetRecord(false)
 
 	for resps := range runCommandGlobally(cmd) {
 		for _, resp := range resps {
@@ -242,43 +242,43 @@ func webVMs(w http.ResponseWriter, r *http.Request) {
 	// we want a map of "hostname + id" to vm info so that it can be sorted
 	infovms := make(map[string]map[string]interface{}, 0)
 
-	for host, vms := range globalVMs(false) {
-		for _, vm := range vms {
-			stateMask := VM_QUIT | VM_ERROR
+	cmdLock.Lock()
+	vms := GlobalVMs()
+	cmdLock.Unlock()
 
-			if vm.GetState()&stateMask != 0 {
-				continue
-			}
+	for _, vm := range vms {
+		stateMask := VM_QUIT | VM_ERROR
 
-			config := vm.Config()
-
-			vmMap := map[string]interface{}{
-				"host": host,
-
-				"id":    vm.GetID(),
-				"name":  vm.GetName(),
-				"state": vm.GetState().String(),
-				"type":  vm.GetType().String(),
-
-				"vcpus":  config.Vcpus,
-				"memory": config.Memory,
-			}
-
-			if config.Networks == nil {
-				vmMap["network"] = make([]int, 0)
-			} else {
-				vmMap["network"] = config.Networks
-			}
-
-			if vm.GetTags() == nil {
-				vmMap["tags"] = make(map[string]string, 0)
-			} else {
-				vmMap["tags"] = vm.GetTags()
-			}
-
-			// The " " is invalid as a hostname, so we use it as a separator.
-			infovms[host+" "+strconv.Itoa(vm.GetID())] = vmMap
+		if vm.GetState()&stateMask != 0 {
+			continue
 		}
+
+		config := vm.Config()
+
+		vmMap := map[string]interface{}{
+			"host":   vm.GetHost(),
+			"id":     vm.GetID(),
+			"name":   vm.GetName(),
+			"state":  vm.GetState().String(),
+			"type":   vm.GetType().String(),
+			"vcpus":  config.Vcpus,
+			"memory": config.Memory,
+		}
+
+		if config.Networks == nil {
+			vmMap["network"] = make([]int, 0)
+		} else {
+			vmMap["network"] = config.Networks
+		}
+
+		if vm.GetTags() == nil {
+			vmMap["tags"] = make(map[string]string, 0)
+		} else {
+			vmMap["tags"] = vm.GetTags()
+		}
+
+		// The " " is invalid as a hostname, so we use it as a separator.
+		infovms[vm.GetHost()+" "+strconv.Itoa(vm.GetID())] = vmMap
 	}
 
 	// We need to pass it as an array for the JSON generation (so the weird keys don't show up)
